+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - helm.v3.ChartOpts: Add KubeVersion field that can be passed to avoid asking the kubernetes API server for the version (https://github.com/pulumi/pulumi-kubernetes/pull/2593)
+- Fix for Helm Import regression (https://github.com/pulumi/pulumi-kubernetes/pull/2605)
 
 ## 4.4.0 (October 12, 2023)